+++ www/test/angularjs-surface.test.js
@@ -21,22 +21,16 @@ const allowedAngularFiles = [
     "public/js/ng-showdown.js",
     "public/js/showdown.min.js",
     "views/builder/index.ejs",
-    "views/items/display.ejs",
     "views/items/index.ejs",
     "views/items/modify.ejs",
     "views/login.ejs",
-    "views/mobs/display.ejs",
     "views/mobs/modify.ejs",
-    "views/quests/display.ejs",
     "views/quests/modify.ejs",
     "views/shared/columnsModal.ejs",
-    "views/shared/displayChangelog.ejs",
-    "views/shared/displayChangelogEdit.ejs",
     "views/shared/filtersModal.ejs",
     "views/shared/markdown.ejs",
     "views/shared/mobModal.ejs",
     "views/shared/questModal.ejs",
-    "views/wiki/display.ejs",
     "views/wiki/modify.ejs"
 ];